refactor(tests): split chat polish coverage
