+++ scripts/run-tests.mjs
@@ -320,7 +320,10 @@ export const SUITES = {
     "src/lib/code-block-collapse.test.ts",
     "src/components/message-bubble-jump-to-line.test.ts",
     "src/components/message-bubble-file-links.test.ts",
-    "src/components/chat-view-polish.test.ts",
+    "src/components/chat-view-polish-tools-activity.test.ts",
+    "src/components/chat-view-polish-header-composer.test.ts",
+    "src/components/chat-view-polish-attachments-mentions.test.ts",
+    "src/components/chat-view-polish-edit-review.test.ts",
     "src/components/chat-composer-footer-band.test.ts",
     "src/components/composer-runtime-chip.test.ts",
     "src/components/composer-git-chip.test.ts",
