+++ src/components/chat-view-polish-header-composer.test.ts
@@ -1,221 +1,23 @@
 // @ts-nocheck
 import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
-
-const source = readFileSync(new URL("./chat-view.tsx", import.meta.url), "utf8");
-// The empty state (the familiar's starting page) was extracted when it became
-// task-aware; its launch-screen pins now read the dedicated file.
-const emptyStateSource = readFileSync(new URL("./chat-empty-state.tsx", import.meta.url), "utf8");
-const styles = ["cave-md", "cave-composer", "chat-list", "calendar", "cave-chat"]
-  .map((sheet) => readFileSync(new URL(`../styles/${sheet}.css`, import.meta.url), "utf8"))
-  .join("\n");
-const globalsSrc = readFileSync(new URL("../app/globals.css", import.meta.url), "utf8");
-// fileToAttachment moved to the shared lib (reused by the home composer).
-const attachmentsLib = readFileSync(new URL("../lib/chat-attachments.ts", import.meta.url), "utf8");
-// The staging state machine (cap, drag overlay, paste) lives in the shared hook.
-const attachStagingHook = readFileSync(new URL("../lib/use-attachment-staging.ts", import.meta.url), "utf8");
-// The inline slash menus (options, dismiss, keyboard branches) live in theirs.
-const menusHookSource = readFileSync(new URL("../lib/use-inline-slash-menus.ts", import.meta.url), "utf8");
+import {
+  attachmentsLib,
+  attachStagingHook,
+  emptyStateSource,
+  globalsSrc,
+  menusHookSource,
+  source,
+  splitReasoning,
+  styles,
+  turnRow,
+} from "./chat-view-polish-fixtures.ts";
 
 assert.doesNotMatch(
   globalsSrc,
   /\.cave-code-page[\s\S]*?cave-composer|data-code-inline-toolbar|code-mode-toggle/,
   "Retired Code page and inline toolbar styling should stay removed",
 );
-const turnRow = source.match(/function TurnRowImpl[\s\S]*?\n}\n\ntype TurnRowProps/)?.[0] ?? "";
-const splitReasoning = source.match(/function splitReasoning[\s\S]*?\n}\n\n\/\/ ── ChatEmptyState/)?.[0] ?? "";
-
-assert.match(
-  source,
-  /fetch\("\/api\/chat\/send"[\s\S]*body: JSON\.stringify\(\{[\s\S]*attachments: stripPreviewOnlyAttachmentFieldsKeepingImages\(outgoingAttachments\)/,
-  "Chat send should strip preview-only attachment fields before POSTing, keeping image payloads so the harness can see them",
-);
-
-assert.match(
-  attachmentsLib,
-  /if \(file\.size > MAX_ATTACHMENT_IMAGE_BYTES\) \{[\s\S]*?attachment\.truncated = true;/,
-  "Oversized image attachments should be capped at capture time and marked like truncated text",
-);
-
-assert.match(
-  source,
-  /const isImage = \(attachment\.mimeType \?\? attachment\.type\)\?\.startsWith\("image\/"\)/,
-  "Attachment lightbox should fall back to legacy attachment.type for images",
-);
-
-assert.match(
-  source,
-  /role="dialog"[\s\S]*aria-modal="true"/,
-  "Attachment lightbox should expose modal dialog semantics",
-);
-
-assert.match(
-  splitReasoning,
-  /tagRe\.exec\(text\)/,
-  "Reasoning splitting should use a streaming-safe tag scanner",
-);
-
-assert.match(
-  splitReasoning,
-  /if \(activeTag\) \{[\s\S]*reasoningParts\.push\(text\.slice\(reasoningStart\)\.trim\(\)\)/,
-  "Unclosed reasoning blocks should be captured instead of leaking raw tags into chat",
-);
-
-assert.match(
-  splitReasoning,
-  /if \(!activeTag && closing\) \{[\s\S]*cursor = tagRe\.lastIndex/,
-  "Unmatched closing reasoning tags should be hidden instead of leaking raw markup into chat",
-);
-
-assert.match(
-  turnRow,
-  /<ToolGroup|<ReasoningBlock/,
-  "Assistant turns should render tool-use and reasoning chrome in collapsed transcript blocks",
-);
-
-assert.match(
-  turnRow,
-  /const reasoningSplit = splitReasoning\(extractAgentAttachmentMarkers\(turn\.text\)\.text\)[\s\S]*const inlineReasoning = reasoningSplit\.reasoning[\s\S]*const \{ visible: visibleWithGh, suggestions: nextPaths \} = extractNextPaths\(skillSplit\.visible\)/,
-  "Assistant turns should split visible content from collapsible reasoning before extracting next-path suggestions",
-);
-
-assert.match(
-  source,
-  /function ReasoningBlock[\s\S]*<details[\s\S]*data-default-collapsed="true"[\s\S]*Thinking[\s\S]*<RichText text=\{reasoning\}/,
-  "ReasoningBlock should render thinking in a collapsed disclosure with formatted text",
-);
-
-// Thinking is togglable: the global Show-thinking preference opens every
-// reasoning block at once via a controlled `open` (default-collapsed in markup).
-assert.match(
-  source,
-  /function ReasoningBlock[\s\S]*const \[showThinking\] = useShowThinking\(\)[\s\S]*open=\{showThinking \|\| undefined\}/,
-  "ReasoningBlock open state is driven by the global Show-thinking preference",
-);
-assert.match(
-  source,
-  /function SessionOverflowMenu[\s\S]*useShowThinking\(\)[\s\S]*checked=\{showThinking\}[\s\S]*\{showThinking \? "Hide thinking" : "Show thinking"\}/,
-  "The session overflow menu carries the global Show-thinking toggle",
-);
-
-assert.match(
-  source,
-  /function ToolGroup[\s\S]*<details[\s\S]*data-default-collapsed="true"[\s\S]*Tool activity[\s\S]*tools\.map[\s\S]*<ToolBlock/,
-  "ToolGroup should render tool calls in a collapsed disclosure",
-);
-
-assert.match(
-  source,
-  /function ToolBlock[\s\S]*<details[\s\S]*data-default-collapsed="true"[\s\S]*<summary[\s\S]*tool\.name[\s\S]*<ToolInputView input=\{tool\.input\}[\s\S]*<SyntaxBlock text=\{prettyToolOutput\(tool\.output\)\}/,
-  "ToolBlock keeps payloads collapsed, renders readable input fields, and pretty-prints output",
-);
-
-// JSON tool input is converted to a human-readable labelled field list, with
-// the raw JSON available behind a toggle.
-assert.match(
-  source,
-  /function ToolInputView[\s\S]*toolReadableFields\(input\)[\s\S]*showRaw \? <SyntaxBlock text=\{input\}/,
-  "ToolInputView renders readable fields by default and raw JSON on toggle",
-);
-assert.match(
-  source,
-  /function ToolFieldList[\s\S]*field\.label[\s\S]*field\.value/,
-  "ToolFieldList renders each readable field's humanised label and value",
-);
-
-// Tool rows are color-coded by category for quick visual inspection.
-assert.match(source, /import \{ toolVisual \} from "@\/lib\/tool-visual"/, "chat view imports the tool visual map");
-assert.match(
-  source,
-  /function ToolBlock[\s\S]*const visual = toolVisual\(tool\.name\)[\s\S]*data-tool-category=\{visual\.category\}[\s\S]*<Icon name=\{visual\.icon\}/,
-  "ToolBlock should color-code by tool category (data-tool-category + per-category icon)",
-);
-
-// Tool-use disclosures must never default open (the transcript stays clean).
-// ReasoningBlock is the one exception — its `open` is a controlled binding to
-// the global Show-thinking preference (asserted above), not a hardcoded default.
-assert.doesNotMatch(
-  [
-    source.match(/function ToolGroup[\s\S]*?function ToolBlock/)?.[0] ?? "",
-    source.match(/function ToolBlock[\s\S]*?function ToolInputView/)?.[0] ?? "",
-  ].join("\n"),
-  /<details[^>]*\sopen(?:=|\s|>)/,
-  "Tool-use disclosures must not default open",
-);
-// A hardcoded `open` (open with no binding) on the reasoning block would defeat
-// the toggle — only the controlled `open={showThinking || undefined}` is allowed.
-assert.doesNotMatch(
-  source.match(/function ReasoningBlock[\s\S]*?function ProgressGroup/)?.[0] ?? "",
-  /<details[^>]*\sopen(?:\s|>)/,
-  "ReasoningBlock must not hardcode the disclosure open",
-);
-
-// --- Tool activity renders in a designated section on settled turns ---
-
-// No per-turn show/hide toggle: the designated section is always present
-// (collapsed) instead, so prose and tool usage are cleanly separated.
-assert.doesNotMatch(
-  turnRow,
-  /showTools|showToolsOverride|cave-turn-tools-toggle/,
-  "the settled-turn tool show/hide toggle is gone — tools live in a designated section",
-);
-
-assert.match(
-  turnRow,
-  /segments=\{renderSegments\}/,
-  "MessageBubble renders the artifact-aware renderSegments",
-);
-
-assert.match(
-  turnRow,
-  /renderSegments = split\.some\(\(s\) => s\.kind === "block"\) \? split : undefined/,
-  "settled turns render prose (+ artifacts) only — tool blocks are not woven into the text",
-);
-
-assert.match(
-  turnRow,
-  /!turn\.pending && turn\.tools\?\.length/,
-  "settled turns that used tools render a designated tool section",
-);
-assert.match(
-  turnRow,
-  /cave-edit-cards[\s\S]*editCards\.map\(\(tool\) => <ToolBlock/,
-  "edit-tool cards stay visible inline on settled turns (not buried in the collapsed rollup)",
-);
-assert.match(
-  turnRow,
-  /const isEditCard = \(t: ToolEvent\) =>\s*toolInputAsDiff\(t\.name, t\.input\) != null;/,
-  "any structured file mutation diff stays visible inline, even when the tool input only has a relative path",
-);
-// Golden path 4 (cave-qva4): a multi-file turn gets ONE aggregate entry into
-// the working-tree review, riding the per-card cave:open-file-diff contract.
-assert.match(
-  turnRow,
-  /const editedFiles = Array\.from\(\s*\n\s*new Set\(\s*\n\s*editCards\s*\n\s*\.map\(\(t\) => toolTargetFile\(t\.name, t\.input\)\)/,
-  "the aggregate counts DISTINCT edited files (the same file edited twice is one change)",
-);
-assert.match(
-  turnRow,
-  /\{editedFiles\.length > 1 \? \([\s\S]{0,400}?\{editedFiles\.length\} files changed/,
-  "turns that edited more than one distinct file render the 'N files changed' chip (single-file turns keep just the card's own Review)",
-);
-assert.match(
-  turnRow,
-  /aria-label=\{`Review all \$\{editedFiles\.length\} changed files in the Changes tab`\}[\s\S]{0,300}?cave:open-file-diff/,
-  "Review all opens the Changes tab through the cards' existing event contract",
-);
-assert.match(
-  turnRow,
-  /otherTools\.length \? <ToolGroup tools=\{otherTools\}/,
-  "non-edit tool activity still collapses into the designated ToolGroup",
-);
-
-assert.match(
-  turnRow,
-  /<MessageBubble[\s\S]*role="assistant"[\s\S]*content=\{visible \|\| \(turn\.pending \? "…" : ""\)\}/,
-  "Assistant turns should render only filtered visible content",
-);
-
 assert.match(
   source,
   /<header className="cave-chat-linear-header reveal-scope"/,
@@ -920,392 +722,4 @@ assert.match(
 
 // — CHAT-D1-04: @-mention repo files in the composer —
 
-// Behavioral: mention token parsing + fuzzy ranking (src/lib/file-mention.ts)
-const { fileMentionToken, filterFileMentions, MAX_FILE_MENTIONS, FILE_MENTION_RESULT_LIMIT } =
-  await import("../lib/file-mention.ts");
 
-assert.deepEqual(
-  fileMentionToken("@", 1),
-  { start: 0, query: "" },
-  "A bare `@` at the start of the composer opens an empty-query mention token",
-);
-assert.deepEqual(
-  fileMentionToken("look at @src/ch", 15),
-  { start: 8, query: "src/ch" },
-  "`@` after whitespace yields the text between the @ and the caret (slashes allowed)",
-);
-assert.equal(
-  fileMentionToken("mail me a@b.com", 15),
-  null,
-  "Mid-word `@` (emails) must not open the picker — the @ must start the text or follow whitespace",
-);
-assert.equal(
-  fileMentionToken("@src foo", 8),
-  null,
-  "Whitespace between the @ and the caret closes the token",
-);
-assert.equal(
-  fileMentionToken("@a@b", 4),
-  null,
-  "A second `@` inside the query invalidates the token",
-);
-assert.deepEqual(
-  fileMentionToken("@src/ch trailing", 7),
-  { start: 0, query: "src/ch" },
-  "Only text up to the caret counts as the query",
-);
-assert.equal(
-  fileMentionToken("/help", 5),
-  null,
-  "A `/` first token is never a mention — slash menu and mention menu stay disjoint",
-);
-
-const mentionIndexFixture = [
-  "src/components/chat-view.tsx",
-  "src/lib/chat-attachments.ts",
-  "src/lib/file-mention.ts",
-  "docs/changelog.md",
-  "chat.ts",
-];
-assert.deepEqual(
-  filterFileMentions(mentionIndexFixture, "chat")[0],
-  "chat.ts",
-  "Basename-prefix matches rank above basename/path substring matches",
-);
-assert.ok(
-  filterFileMentions(mentionIndexFixture, "chat").includes("src/components/chat-view.tsx"),
-  "Basename substring matches are included after prefix matches",
-);
-assert.deepEqual(
-  filterFileMentions(mentionIndexFixture, "scmvt"),
-  ["src/components/chat-view.tsx"],
-  "Subsequence matching catches scattered-character queries",
-);
-assert.deepEqual(
-  filterFileMentions(mentionIndexFixture, "zzz"),
-  [],
-  "Non-matching queries return no rows",
-);
-assert.equal(
-  filterFileMentions(mentionIndexFixture, "", 2).length,
-  2,
-  "The result limit caps the list (empty query returns the head of the index)",
-);
-assert.equal(MAX_FILE_MENTIONS, 10, "Mentions are capped at 10 per send");
-assert.ok(FILE_MENTION_RESULT_LIMIT <= 15, "The picker shows a short list (~12), not the whole index");
-
-// Pins: file-index route mirrors the /api/changes security posture
-const filesRouteSource = readFileSync(
-  new URL("../app/api/project/files/route.ts", import.meta.url),
-  "utf8",
-);
-assert.match(
-  filesRouteSource,
-  /execFileAsync\("git", args, \{/,
-  "/api/project/files must run git through execFile with an argument array (no shell)",
-);
-assert.doesNotMatch(
-  filesRouteSource,
-  /\bexec\(|shell:\s*true|spawnSync\(/,
-  "/api/project/files must never interpolate the root into a shell command",
-);
-assert.match(
-  filesRouteSource,
-  /if \(!path\.isAbsolute\(root\)\)/,
-  "/api/project/files must reject relative roots",
-);
-assert.match(
-  filesRouteSource,
-  /import \{ resolveAllowedProjectPath \} from "@\/lib\/server\/project-paths"/,
-  "/api/project/files must reuse the standard allowed-root guard",
-);
-assert.match(
-  filesRouteSource,
-  /const allowedRoot = resolveAllowedProjectPath\(root\);[\s\S]*?if \(!allowedRoot\)[\s\S]*?path not allowed[\s\S]*?status: 403/,
-  "/api/project/files must reject roots outside allowed workspaces",
-);
-assert.match(
-  filesRouteSource,
-  /try \{[\s\S]*?fs\.realpathSync\(allowedRoot\);[\s\S]*?fs\.statSync\(real\);[\s\S]*?\} catch/,
-  "/api/project/files must realpath and stat the allowed root inside a guarded block",
-);
-assert.match(
-  filesRouteSource,
-  /\{ ok: true, repo: false, error: resolved\.error \}/,
-  "Not-a-repo must be a distinct non-error state, not a 4xx/5xx",
-);
-assert.match(
-  filesRouteSource,
-  /"ls-files",\s*\n\s*"-z",\s*\n\s*"--cached",\s*\n\s*"--others",\s*\n\s*"--exclude-standard",/,
-  "The index must list tracked plus untracked-but-not-ignored files, NUL-separated",
-);
-assert.match(
-  filesRouteSource,
-  /const MAX_FILES = 5000;/,
-  "The index must cap at ~5000 paths",
-);
-assert.match(
-  filesRouteSource,
-  /truncated = all\.length > MAX_FILES/,
-  "An over-cap index must set the truncated flag",
-);
-assert.match(
-  filesRouteSource,
-  /const CACHE_TTL_MS = 10_000;[\s\S]*Date\.now\(\) - cached\.at < CACHE_TTL_MS/,
-  "The route must keep a ~10s module-level cache keyed by root",
-);
-
-// Pins: composer mention picker (chat-view.tsx)
-const mentionSource = readFileSync(new URL("./chat-view.tsx", import.meta.url), "utf8");
-assert.match(
-  mentionSource,
-  /const mentionRoot = activeProjectRoot\.trim\(\);/,
-  "The mention root must use the selected project root",
-);
-assert.match(
-  mentionSource,
-  /projectRoot: requestProjectRoot/,
-  "The send body must use the vetted project root (selected project, minus unregistered session-cwd echoes)",
-);
-assert.match(
-  mentionSource,
-  /new URLSearchParams\(\{ root: mentionRoot, familiarId: familiar\.id \}\)/,
-  "The picker must fetch the file index for the chat's project root scoped to the active familiar",
-);
-assert.match(
-  mentionSource,
-  /id=\{mentionListboxId\} role="listbox" aria-label="Workspace files"/,
-  "The mention popover must be a listbox (ARIA parity with the slash menu, #423)",
-);
-assert.match(
-  mentionSource,
-  /role="option"\s*\n\s*id=\{`\$\{mentionListboxId\}-opt-\$\{i\}`\}\s*\n\s*aria-selected=\{active\}/,
-  "Mention rows must be aria-selected options with stable ids for aria-activedescendant",
-);
-assert.match(
-  mentionSource,
-  /const mentionActiveIdx = mentionOpen \? Math\.min\(mentionIdx, mentionMatches\.length - 1\) : 0;/,
-  "Mention active index must clamp to the current match count",
-);
-assert.match(
-  mentionSource,
-  /setMentionIdx\(\(i\) => \(mentionMatches\.length === 0 \? 0 : Math\.min\(i, mentionMatches\.length - 1\)\)\);/,
-  "Mention index should be brought back in range when the match list shrinks",
-);
-assert.match(
-  mentionSource,
-  /const mentionAriaOverrides: React\.AriaAttributes = mentionOpen\s*\n\s*\? \{\s*\n\s*"aria-expanded": true,\s*\n\s*"aria-controls": mentionListboxId,\s*\n\s*"aria-activedescendant": `\$\{mentionListboxId\}-opt-\$\{mentionActiveIdx\}`,/,
-  "While the mention picker is open it must override the combobox ARIA with the clamped active option",
-);
-assert.match(
-  mentionSource,
-  /\{\.\.\.mentionAriaOverrides\}/,
-  "The composer textarea must apply the mention ARIA overrides after the slash wiring (later JSX attributes win)",
-);
-assert.match(
-  mentionSource,
-  /const active = i === mentionActiveIdx;/,
-  "Mention row highlight should use the same clamped active index as aria-activedescendant",
-);
-
-// Esc precedence (#402): mention dismiss → slash dismiss → busy cancel.
-const mentionComposerKey = mentionSource.match(/const onComposerKey = [\s\S]*?\n  \};/)?.[0] ?? "";
-assert.match(
-  mentionComposerKey,
-  /if \(mentionOpen\) \{[\s\S]*?setMentionDismissed\(true\)/,
-  "Esc with the mention picker open must dismiss the picker",
-);
-// The slash branches live behind the shared hook's dispatcher, so the #402
-// ordering contract anchors on the dispatch call: mention branch → menu
-// dispatcher (consumes Esc while any menu is open) → busy-cancel.
-assert.ok(
-  mentionComposerKey.indexOf("if (mentionOpen) {") <
-    mentionComposerKey.indexOf("handleMenuKey(e)"),
-  "The mention branch must run before the slash-menu dispatcher in onComposerKey",
-);
-assert.ok(
-  mentionComposerKey.indexOf("setMentionDismissed(true)") <
-    mentionComposerKey.indexOf("handleMenuKey(e)") &&
-    mentionComposerKey.indexOf("handleMenuKey(e)") <
-      mentionComposerKey.indexOf("cancelSend()"),
-  "Esc precedence: mention dismiss before slash dismiss (inside the dispatcher) before busy-cancel",
-);
-assert.match(
-  mentionComposerKey,
-  /if \(e\.key === "Tab" \|\| \(e\.key === "Enter" && !e\.shiftKey\)\) \{[\s\S]*?selectMention\(file\)/,
-  "Enter/Tab must insert the highlighted file, never send the draft, while the picker is open",
-);
-assert.match(
-  mentionSource,
-  /setMentionIdx\(0\);\s*\n\s*setMentionDismissed\(false\);/,
-  "Editing the input must re-arm a dismissed mention picker",
-);
-
-// Selection semantics: inline `@path` token + mentionedFiles in the send body.
-assert.match(
-  mentionSource,
-  /const insert = `@\$\{relPath\} `;[\s\S]*?input\.slice\(0, mentionToken\.start\) \+ insert \+ input\.slice\(composerCaret\)/,
-  "Selecting a file must replace the `@query` token with the relative path inline (Claude Code convention)",
-);
-assert.match(
-  mentionSource,
-  /\.\.\.\(outgoingMentions\.length && mentionRoot\s*\n\s*\? \{\s*\n\s*mentionedFiles: outgoingMentions\.slice\(0, MAX_FILE_MENTIONS\),\s*\n\s*mentionedFilesRoot: mentionRoot,/,
-  "The send body must carry mentionedFiles plus the root they are relative to",
-);
-assert.match(
-  mentionSource,
-  /mentionedFiles\s*\n?\s*\.filter\(\(p\) => text\.includes\(`@\$\{p\}`\)\)/,
-  "Only mentions whose @path token survived editing may ride the send",
-);
-assert.match(
-  mentionSource,
-  /setInput\(""\);[\s\S]{0,400}?clearAttachments\(\);\s*\n\s*setMentionedFiles\(\[\]\);/,
-  "Sending must clear staged mentions with the composer",
-);
-
-// Pins: send route validates mentions and appends the prompt block.
-const mentionSendSource = readFileSync(
-  new URL("../app/api/chat/send/route.ts", import.meta.url),
-  "utf8",
-);
-assert.match(
-  mentionSendSource,
-  /async function resolveMentionedFiles\([\s\S]*?relPaths\.slice\(0, MAX_MENTIONED_FILES\)[\s\S]*?\.includes\("\.\."\)[\s\S]*?candidate\.startsWith\(realRoot \+ path\.sep\)/,
-  "/chat/send must validate each mention: cap, repo-relative only, no `..`, prefix containment under the realpathed root",
-);
-assert.match(
-  mentionSendSource,
-  /async function resolveFamiliarWorkspace\([\s\S]*?readFamiliarWorkspaces\(\)[\s\S]*?path\.resolve\(familiarsRoot, familiarId\)[\s\S]*?path\.relative\(familiarsRoot, candidate\)[\s\S]*?relative\.startsWith\("\.\."\)/,
-  "/chat/send must validate default familiar workspace paths under the familiar root while preserving configured workspaces",
-);
-assert.match(
-  mentionSendSource,
-  /const real = await realpath\(candidate\);[\s\S]*?real\.startsWith\(realRoot \+ path\.sep\)/,
-  "/chat/send must re-check containment on the realpathed file so in-repo symlinks cannot smuggle outside paths",
-);
-assert.match(
-  mentionSendSource,
-  /"Referenced files \(open with the Read tool\):",\s*\n\s*\.\.\.absPaths\.map\(\(p\) => `- \$\{p\}`\)/,
-  "Validated mentions must render as the compact Referenced-files prompt block of absolute paths",
-);
-assert.match(
-  mentionSendSource,
-  /appendMentionedFilesBlock\(\s*\n\s*buildPromptWithResponseControls\(\s*\n\s*buildPromptWithAttachments\(/,
-  "The mention block must join the prompt after attachments and response controls are applied",
-);
-assert.match(
-  mentionSendSource,
-  /const resolvedFamiliarWorkspace = !sshRuntime\s*\n\s*\? await resolveFamiliarWorkspace\(body\.familiarId\)\s*\n\s*: undefined;/,
-  "Mention roots must come from the validated familiar workspace, not a client-supplied path",
-);
-assert.match(
-  mentionSendSource,
-  /const mentionedFiles = imagesSupported\s*\n\s*\? await resolveMentionedFiles\(\s*\n\s*body\.mentionedFiles,\s*\n\s*resolvedFamiliarWorkspace,/,
-  "Mentions are only delivered to harnesses that can Read this machine's filesystem, against the validated familiar workspace",
-);
-
-// The top suggested follow-up is flagged as the recommendation (green pulsing
-// border + leading dot), so the most useful next step stands out.
-assert.match(
-  source,
-  /cave-next-path--recommended/,
-  "the first follow-up is marked as the recommended next step",
-);
-
-// Suggestion pills lay out in UNIFORM rows keyed off the chip count: 2 and 4
-// chips pair into two columns (4 = 2×2, never a 3+1 orphan wrap); every other
-// count — legacy 3-chip transcripts included — stacks full-width (cave-wrso,
-// cave-98bs).
-assert.match(
-  source,
-  /className="cave-next-paths" data-count=\{nextPaths\.length\}/,
-  "the chip row stamps its count so CSS can key columns off it",
-);
-assert.match(
-  globalsSrc,
-  /\.cave-next-paths\[data-count="2"\],\s*\n\s*\.cave-next-paths\[data-count="4"\] \{ grid-template-columns: repeat\(2, 1fr\); \}/,
-  "2 and 4 chips pair into two columns (4 renders 2×2)",
-);
-assert.ok(
-  !/\.cave-next-paths\[data-count="3"\]/.test(globalsSrc),
-  "no exactly-3 column rule: the chatturn container (46rem reading column, ~672px inner) can never reach a width where three chips fit, so legacy 3-chip rows stack (cave-98bs)",
-);
-assert.ok(
-  !/\.cave-next-paths \{ grid-template-columns: repeat\(/.test(globalsSrc),
-  "no count-blind multi-column rule survives (it produced 3+1 orphan wraps)",
-);
-
-// File picker resets its value synchronously so re-selecting the same file (or
-// re-attaching after the CSV / 10-cap early returns) still fires onChange.
-assert.ok(
-  source.includes("const files = e.currentTarget.files ? Array.from(e.currentTarget.files) : null;"),
-  "file input snapshots files before reset",
-);
-assert.ok(
-  source.includes('e.currentTarget.value = "";') && !source.includes('fileInputRef.current.value = ""'),
-  "file input resets value synchronously in onChange, not after the async attach",
-);
-
-// Codex inline file-edit card: Edit/Write/MultiEdit/NotebookEdit tool calls
-// render as a visible details card in the transcript. The collapsed summary
-// shows when/status + what file changed; expanding the same card shows the
-// actual diff, matching the Bash/tool-use disclosure pattern.
-assert.match(source, /cave-edit-card/, "mutation tools render as an inline Codex edit card");
-assert.match(source, /diffStat/, "edit card derives a +/- stat");
-assert.match(source, /Review/, "edit card has a Review action");
-assert.match(globalsSrc, /\.cave-edit-card/, "edit card styling exists");
-
-// Review adapts to where the edit can actually be reviewed: a file under the
-// session's project root jumps to the code rail's Changes diff; anything else
-// (familiar-workspace docs, repo-less sessions, relative paths) opens an
-// in-chat modal with this edit's diff instead of dispatching an event nothing
-// can service. The actions row renders on every edit card — not only when an
-// absolute target path exists — so Review is always available.
-assert.match(
-  source,
-  /if \(relPath && targetFile\) \{[\s\S]{0,200}cave:open-file-diff[\s\S]{0,200}setReviewOpen\(true\)/,
-  "Review falls back to the in-chat diff modal when the Changes panel can't show the file",
-);
-assert.match(
-  source,
-  /<Modal[\s\S]{0,200}open=\{reviewOpen\}[\s\S]{0,600}<SyntaxBlock text=\{diff\} lang="diff" \/>/,
-  "the review modal renders this edit's structured diff",
-);
-assert.match(
-  source,
-  /<EditCardActions targetFile=\{targetFile\} diff=\{inputDiff \?\? ""\} displayPath=\{displayPath\} \/>/,
-  "edit-card actions render unconditionally (Review works without an absolute target path)",
-);
-assert.match(globalsSrc, /\.cave-review-modal/, "review modal styling exists");
-assert.match(
-  source,
-  /if \(isEditTool\) \{[\s\S]*<details className="cave-tool-block cave-edit-card"[\s\S]*Edited \{base\}[\s\S]*<DurationText durationMs=\{tool\.durationMs\} \/>[\s\S]*Code changes[\s\S]*<SyntaxBlock text=\{inputDiff\} lang="diff" \/>[\s\S]*<\/details>/,
-  "edit cards should use the same expandable tool details pattern and include the code diff in chat",
-);
-
-// Inline "Undo" reverts the edited file to its last committed state via the
-// changes revert API, resolving the repo-relative path through a context, and
-// pings the Changes panel to refresh.
-assert.match(source, /cave-edit-card__undo/, "edit card has an Undo action");
-assert.match(source, /ToolProjectRootContext/, "edit card resolves project root via context for revert");
-assert.match(source, /"\/api\/changes"/, "Undo posts to the changes revert API");
-assert.match(source, /cave:changes-refresh/, "Undo notifies the changes panel to refresh");
-assert.match(globalsSrc, /\.cave-edit-card__undo/, "Undo button styling exists");
-
-// cave-zvr: composer send hygiene + picker Escape.
-// (3) send() clears the persisted draft synchronously — the 250ms debounced
-//     writer is cancelled if ChatView unmounts right after send, else the
-//     pre-send text reappears as a draft on return.
-assert.match(source, /setInput\(""\);\s*\n\s*\/\/[\s\S]*?clearDraft\(\);/, "send clears the persisted composer draft synchronously");
-// (2) send() resets the enhance strip so it doesn't linger over an empty
-//     composer and let Revert repopulate the already-sent message.
-assert.match(source, /clearDraft\(\);[\s\S]{0,400}?promptEnhance\.reset\(\);/, "send resets the enhance strip state");
-// (1) the slash, /model, /skill and /prompt pickers all dismiss on Escape
-//     (their footers advertise "esc cancel"); previously Esc fell through and
-//     cancelled a live stream. The shared hook guards ONE Escape branch on
-//     menuOpen — the union of all four pickers — so none can leak Esc through.
-assert.match(
-  menusHookSource,
-  /if \(e\.key === "Escape" && menuOpen\) \{\s*\n\s*e\.preventDefault\(\);\s*\n\s*setSlashDismissed\(true\);\s*\n\s*return true;\s*\n\s*\}/,
-  "the slash, model, skill and prompt pickers all dismiss on Escape (setSlashDismissed behind menuOpen)",
-);
